requirements.txt
torch
torchvision
scikit-learn
pandas
numpy
matplotlib
Pillow
PyYAML
tqdm
jupyter
ipykernel
